refactor(core): move LogEvent/OnLogEvent into a core log-sink port
infra needs the generic log-event callback but may not depend on app
(AGENTS.md dependency rule), so the shape moves to core/ports/log-sink.
packages/app/src/update re-exports it, keeping the app public surface
unchanged. No behavior change.
