+++ packages/app/src/update/index.ts
@@ -4,4 +4,12 @@
  */
 export * from './auto-update';
 export * from './check-updates';
-export * from './log-event';
+/**
+ * `LogEvent`/`OnLogEvent` now live in `core` (`ports/log-sink`) because
+ * `infra` needs them too and may not depend on `app`. Re-exported here so
+ * the `app` public surface stays unchanged for delivery layers.
+ */
+export type {
+  LogEvent,
+  OnLogEvent,
+} from '@ai-primitives-hub/core';
